# table-reservation-backend/src/main/java/com/example/tablereservation/dto/auth/LoginResponse.java
package com.example.tablereservation.dto.auth;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class LoginResponse {
    private String token;
    private String type = "Bearer";
    private String firstName;
    private String lastName;
    private String email;
}